Begin converting serverside representations of money to cents (as int)

+++ server/db/models/Payment.js
@@ -5,7 +5,7 @@ module.exports = {
   attributes: {
 
     amount: {
-      type: Sequelize.DECIMAL(5, 2),
+      type: Sequelize.INTEGER,
       allowNull: false,
     },
